refactor: extract routes into modules
also use idiomatic config for /explore, but no functional change
